+++ php_mapi/mapi.stub.php
@@ -113,3 +113,4 @@ function nsp_getuserinfo(string $username) : array|false {}
 function nsp_setuserpasswd(string $username, string $oldpass, string $newpass) : bool {}
 function mapi_linkmessage(resource $ses, ?string $srcheid = null, ?string $msgeid = null) : mixed {}
 function mapi_ianatz_to_tzdef(string $tz) : string|false {}
+function mapi_strerror(int $code) : string {}